pyproject.toml
[project]
name = "innovation-factory"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.2.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.12",
    "apx==0.2.6",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[tool.apx.metadata]
app-name = "innovation-factory"
app-slug = "innovation_factory"
app-entrypoint = "innovation_factory.backend.app:app"
api-prefix = "/api"
metadata-path = "src/innovation_factory/_metadata.py"

[tool.apx.ui]
root = "src/innovation_factory/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[tool.ty.src]
exclude = ["scripts/", "**/seed_uc_tables.py", "**/create_dashboard.py"]

[tool.ty.rules]
deprecated = "ignore"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require live Databricks connection",
]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.uv]

[[tool.uv.index]]
name = "apx-index"
url = "https://databricks-solutions.github.io/apx/simple"

[tool.uv.sources]

[tool.uv.sources.apx]
index = "apx-index"

[tool.hatch.build.hooks.version]
path = "src/innovation_factory/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/innovation_factory/__dist__", "src/innovation_factory/_metadata.py"]
exclude = ["src/innovation_factory/ui"]
